version avec boucle correction

+++ TP1/fonctions.py
@@ -5,17 +5,17 @@ def puissance(a,b):
 	#return a**b
 	
 
-	if not type(a) is int:
-		if not type(b) is int:
+	if not type(a) is float:
+		if not type(b) is float:
 			a**b
 	else
-			if b == 0:
-				return 1
-			elif b > 0:
-				resultat = a
-				for _ in range(1, b):
-					resultat *= a
-				return resultat
-			else:
-				return 1 / puissance(a, -b)
+		if b == 0:
+			return 1
+		elif b > 0:
+			resultat = a
+			for _ in range(1, b):
+				resultat *= a
+			return resultat
+		else:
+			return 1 / puissance(a, -b)
